requirements.txt
# Minimal runtime deps for v0.3.0 (health check + DB)
sqlalchemy
psycopg[binary]     # psycopg v3 driver
python-dotenv==1.1.1
psycopg2-binary
python-dotenv
